Update yesNoPoll
This enables update mainly focused on increasing efficiency and decreasing contract interaction cost. The voting function was redesigned and now uses the binary values 1 and 2 in order to determine votes instead of using strings.  Also fixed a security flaw that allowed users to cast invalid votes. This fix is also in the vote function and can be found on lines 43-44.